requirements.txt
django
djangorestframework
python-decouple
django-phonenumber-field
django-phonenumbers
djangorestframework-simplejwt
djoser
